requirements.txt
flask
sqlalchemy
pyodbc
python-dotenv
psycopg2
pandas